Filter HRUs by their variable values at harvest/kill

The function can be complementary to plot_variable_at_harvkill() plots. If any issues are identified in the boxplots, the results can be filtered for further analysis with filter_hru_at_harvkill().

Usage

filter_hru_at_harvkill(sim_verify, ...)

Arguments

sim_verify

Simulation output of the function run_swat_verification() as a single object or as a list of objects (representing run_swat_verification() function runs with different settings). To plot the heat units at least the output option outputs = 'mgt' must be set in run_swat_verification().

...

Boolean operations to be applied in the filter operation of the management simulation outputs. Possible variables are crop, phu, plant_bioms, yield, water_stress, aero_stress, temp_stress, n_stress and p_stress.

Examples

if (FALSE) { # \dontrun{
# Filter all HRUs with unusually high PHU values at harvest kill
filter_hru_at_harvkill(sim_verify, phu > 5)

# Filter all HRUs with crop 'wbar' or 'wira' planted but no yield
filter_hru_at_harvkill(sim_verify, crop %in% c('wbar', 'wira'), yield == 0)
} # }