Plot simulated variables of water partition of filtered HRUs saved in hru_wb_aa

Plot simulated variables of water partition of filtered HRUs saved in hru_wb_aa

Usage

plot_water_partition(
  sim_verify,
  tile = NULL,
  lum = NULL,
  mgt = NULL,
  soil = NULL,
  exclude_lum = c("urhd_lum", "urmd_lum", "urml_lum", "urld_lum", "ucom_lum", "uidu_lum",
    "utrn_lum", "uins_lum", "urbn_lum"),
  boxpoints = TRUE
)

Arguments

sim_verify

Simulation output of the function run_swat_verification(). To plot the heat units at least the output option outputs = 'wb' must be set in run_swat_verification()

tile

Optional Boolean TRUE for selecting HRUs with working tiles, FALSE - without working tiles and NULL for selecting all HRUs.

lum

Optional character vector with landuse.lum labels

mgt

Optional character vector with management labels as defined in management.sch.

soil

Optional character vector with soil type labels as defined in the soil data.

exclude_lum

Character vector to define land uses which are excluded in the printed table.

boxpoints

Optional Boolean TRUE for displaying outliers, FALSE for hiding them. Default = TRUE

Value

plotly figure object

Examples

if (FALSE) { # \dontrun{
plot_water_partition(sim_nostress, tile = TRUE)
} # }