Print the average annual qtile for HRUs

print_avannual_qtile prints a table with the average annual qtile in mm for HRUs that used a tile flow parametrization in landuse.lum

Usage

print_avannual_qtile(
  sim_verify,
  exclude_lum = c("urhd_lum", "urmd_lum", "urml_lum", "urld_lum", "ucom_lum", "uidu_lum",
    "utrn_lum", "uins_lum", "urbn_lum")
)

Arguments

sim_verify

Simulation output of the function run_swat_verification(). To plot the heat units at least the output option outputs = 'wb' must be set in run_swat_verification()

exclude_lum

Character vector to define land uses which are excluded in the printed table.

Value

Returns a table with hru ids average annual qtile and attributes.